requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0 